UI Changes
 A user interface (UI) update and new navigation features have been introduced:

Second header: A new information bar has been added below the main header. It displays:
The modification version (on the left).
The option counter displayed as “current / all” (on the right).
The title of the current submenu (center).

Intuitive Navigation (Back): The center title in the Second Header is now fully interactive. Clicking on the menu name acts as a “Back” button, allowing users to quickly return to the previous tab.